Fixed the PageEditor Move Feature
The issues with moving a sensor around the Page being edited has been fixed. This was due to JavaFX not calculating the bounds of the sensor being moved before another drag event fired, resulting in incorrect data being input into the code executed as a result from the drag event firing. Some function calls where added that cause the bounds to be recalculate so this no longer occurs